requirements.txt
# FastAPI Rails-style Boilerplate - Production Dependencies
# Python 3.11+

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Validation & serialization
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Background jobs
celery[redis]==5.4.0
redis==5.2.1

# Environment & CLI
python-dotenv==1.0.1
click==8.1.7

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Utilities
python-multipart==0.0.17
pyyaml==6.0.2
